+++ test/analyzer/lsp_test.cpp
@@ -719,7 +719,6 @@ TEST_F(SemanticTokenTest, encode) {
   this->testEncode(HighlightTokenClass::NUMBER, SemanticTokenTypes::number_, 0);
   this->testEncode(HighlightTokenClass::REGEX, SemanticTokenTypes::regexp_, 0);
   this->testEncode(HighlightTokenClass::STRING, SemanticTokenTypes::string_, 0);
-  this->testEncode(HighlightTokenClass::SIGNAL, SemanticTokenTypes::event_, 0);
   this->testEncode(HighlightTokenClass::COMMAND, SemanticTokenTypes::function_, 0);
   this->testEncode(HighlightTokenClass::COMMAND_ARG, SemanticTokenTypes::parameter_, 0);
   this->testEncode(HighlightTokenClass::REDIRECT, SemanticTokenTypes::operator_, 0);