+++ test/lexer/lexer_test.cpp
@@ -689,28 +689,6 @@ TEST_F(LexerTest_Lv1, invalid_regex2) {
   ASSERT_NO_FATAL_FAILURE(EXPECT(TokenKind::INVALID, "$"));
 }
 
-TEST_F(LexerTest_Lv1, signal1) {
-  const char *text = "%'de1'";
-  this->initLexer(text);
-  ASSERT_NO_FATAL_FAILURE(EXPECT(TokenKind::SIGNAL_LITERAL, text, TokenKind::EOS, ""));
-  ASSERT_NO_FATAL_FAILURE(this->assertLexerMode(yycEXPR));
-}
-
-TEST_F(LexerTest_Lv1, signal2) {
-  const char *text = "%'INT'";
-  this->initLexer(text);
-  ASSERT_NO_FATAL_FAILURE(EXPECT(TokenKind::SIGNAL_LITERAL, text, TokenKind::EOS, ""));
-  ASSERT_NO_FATAL_FAILURE(this->assertLexerMode(yycEXPR));
-}
-
-TEST_F(LexerTest_Lv1, invalid_signal) {
-  const char *text = "%'000'";
-  this->initLexer(text);
-  ASSERT_NO_FATAL_FAILURE(
-      EXPECT(TokenKind::COMMAND, "%", TokenKind::STRING_LITERAL, "'000'", TokenKind::EOS, ""));
-  ASSERT_NO_FATAL_FAILURE(this->assertLexerMode(yycSTMT));
-}
-
 TEST_F(LexerTest_Lv1, subCmd1) {
   const char *text = "$(";
   this->initLexer(text);