+++ CHANGELOG.md
@@ -12,6 +12,7 @@
         - if corresponding signal is not found, ``Signals#[]`` method throw ``KeyNotFoundError``
     - add ``Signal#trap`` method for get and set signal handler corresponding to signal
         - if specified ``SIG_NUL``, does not set signal handler
+- **Breaking Change**: remove signal literal. use signal constants instead
 - now some builtin variables are constants
     - ``ON_ASSERT``, ``ON_ERR``, ``ON_EXIT``
     - ``TRUE``, ``True``, ``true``